Update to new lucene spatial module from lucene 4.0

+++ src/main/java/com/github/uklance/domain/ScoredItem.java
@@ -1,10 +1,10 @@
 package com.github.uklance.domain;
 
 public class ScoredItem implements Comparable<ScoredItem> {
-	private float score;
+	private double score;
 	private Item item;
 	
-	public ScoredItem(Item item, float score) {
+	public ScoredItem(Item item, double score) {
 		super();
 		this.item = item;
 		this.score = score;
@@ -12,8 +12,7 @@ public class ScoredItem implements Comparable<ScoredItem> {
 	public int compareTo(ScoredItem other) {
 		return (score == other.score) ? 0 : ((score < other.score) ? -1 : 1);
 	}
-	
-	public float getScore() {
+	public double getScore() {
 		return score;
 	}
 	public Item getItem() {
